# Introduce extended pixel conversion