--- a/src/ImageSharp/PixelFormats/PixelConversionModifiers.cs
+++ b/src/ImageSharp/PixelFormats/PixelConversionModifiers.cs
@@ -3,14 +3,38 @@
 
 using System;
 
+using SixLabors.ImageSharp.ColorSpaces.Companding;
+
 namespace SixLabors.ImageSharp.PixelFormats
 {
+    /// <summary>
+    /// Flags responsible to select additional operations which could be effitiently applied in
+    /// <see cref="PixelOperations{TPixel}.ToVector4(SixLabors.ImageSharp.Configuration,System.ReadOnlySpan{TPixel},System.Span{System.Numerics.Vector4},SixLabors.ImageSharp.PixelFormats.PixelConversionModifiers)"/>
+    /// or
+    /// <see cref="PixelOperations{TPixel}.FromVector4(SixLabors.ImageSharp.Configuration,System.Span{System.Numerics.Vector4},System.Span{TPixel},SixLabors.ImageSharp.PixelFormats.PixelConversionModifiers)"/>
+    /// knowing the pixel type.
+    /// </summary>
     [Flags]
     internal enum PixelConversionModifiers
     {
+        /// <summary>
+        /// No special operation is selected
+        /// </summary>
         None = 0,
+
+        /// <summary>
+        /// Select <see cref="IPixel.ToScaledVector4"/> <see cref="IPixel.FromScaledVector4"/> instead the standard (non scaled) variants.
+        /// </summary>
         Scale = 1 << 0,
+
+        /// <summary>
+        /// Enable alpha premultiplication / unpremultiplication
+        /// </summary>
         Premultiply = 1 << 1,
+
+        /// <summary>
+        /// Enable SRGB companding (defined in <see cref="SRgbCompanding"/>).
+        /// </summary>
         SRgbCompand = 1 << 2,
     }
 }