… src/ImageSharp/PixelFormats/PixelImplementations/Generated/Rgb24.PixelOperations.Generated.cs
… src/ImageSharp/PixelFormats/PixelImplementations/Generated/_Common.ttinclude
… src/ImageSharp/PixelFormats/PixelImplementations/Rgba32.PixelOperations.cs
… src/ImageSharp/PixelFormats/PixelImplementations/RgbaVector.PixelOperations.cs
… src/ImageSharp/PixelFormats/PixelOperations{TPixel}.cs
… src/ImageSharp/PixelFormats/Utils/Vector4Converters.Default.cs
… src/ImageSharp/PixelFormats/Utils/Vector4Converters.RgbaCompatible.cs
… src/ImageSharp/PixelFormats/Utils/Vector4Converters.cs
… src/ImageSharp/Processing/Processors/Convolution/Convolution2DProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/Convolution2PassProcessor.cs
… src/ImageSharp/Processing/Processors/Convolution/ConvolutionProcessor.cs
… src/ImageSharp/Processing/Processors/Dithering/PaletteDitherProcessorBase.cs
… src/ImageSharp/Processing/Processors/Filters/FilterProcessor.cs
… src/ImageSharp/Processing/Processors/Quantization/FrameQuantizerBase{TPixel}.cs
… src/ImageSharp/Processing/Processors/Transforms/AffineTransformProcessor.cs
… src/ImageSharp/Processing/Processors/Transforms/ProjectiveTransformProcessor.cs
… src/ImageSharp/Processing/Processors/Transforms/Resize/ResizeProcessor.cs
… tests/ImageSharp.Benchmarks/Color/Bulk/FromVector4.cs
… tests/ImageSharp.Benchmarks/PixelBlenders/PorterDuffBulkVsPixel.cs
… tests/ImageSharp.Benchmarks/Samplers/Resize.cs
… tests/ImageSharp.Tests/Colorspaces/Companding/CompandingTests.cs
… tests/ImageSharp.Tests/PixelFormats/PixelOperations/PixelOperationsTests.Bgr24OperationsTests.cs
… tests/ImageSharp.Tests/PixelFormats/PixelOperations/PixelOperationsTests.Rgb24OperationsTests.cs
… tests/ImageSharp.Tests/PixelFormats/PixelOperations/PixelOperationsTests.cs
… tests/ImageSharp.Tests/TestUtilities/TestImageExtensions.cs