@antonfirsov 2019-04-01
I'm merging this now to unblock stuff like #870. Naming tweaks could be added in a follow-up PR.